admin/requirements.txt
Flask==1.1.2
google-cloud-firestore==1.8.1
